finrl_contest/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=6.0

# RL dependencies
torch>=1.10.0
gymnasium>=0.28.0

# LLM and agent dependencies
crewai>=0.28.0
crewai[tools]>=0.28.0
langchain>=0.0.267
openai>=0.27.0

# Financial data and analysis
yfinance>=0.1.70
pandas-ta>=0.3.14b0
ta-lib-binary>=0.4.24
statsmodels>=0.13.0
scikit-learn>=1.0.0

# Testing
pytest>=7.0.0
